services/auth/requirements.txt
fastapi
uvicorn[standard]
bcrypt
psycopg2-binary
